feat: clean and add encoding test
