+++ src/db/query-validation.ts
@@ -1,32 +1,11 @@
-import { IndexHeader, IndexMeta } from "../index-file/meta";
-import {
-  FieldType,
-  OrderBy,
-  Query,
-  Schema,
-  SelectField,
-  WhereNode,
-} from "./database";
+import { IndexHeader } from "../file/meta";
+import { FieldType } from "./database";
+import { OrderBy, Schema, Query, SelectField, WhereNode } from "./query-lang";
 
-/**
- * containsType checks if the given compositeType includes the specified singleType.
- * It uses a bitwise AND operation to determine if the singleType's bit is set in the compositeType.
- *
- * @param {bigint} compositeType - A bigint representing a composite of multiple types.
- * @param {FieldType} singleType - The specific type to check for within the compositeType.
- * @returns {boolean} - Returns true if singleType is included in compositeType, false otherwise.
- */
 function checkType(headerType: number[], queryType: FieldType): boolean {
   return headerType.includes(queryType);
 }
 
-/**
- * validateWhere validates the 'where' clause of the query.
- *
- * @param {WhereNode<T>[] | undefined} where - The 'where' clause to validate.
- * @param {IndexMeta[]} headers - List of headers to check field existence and type compatibility.
- * @throws {Error} - Throws an error if the 'where' clause is missing, invalid, or refers to non-existent fields.
- */
 function validateWhere<T extends Schema>(
   where: WhereNode<T>[] | undefined,
   headers: IndexHeader[],
@@ -104,14 +83,6 @@ function validateWhere<T extends Schema>(
   }
 }
 
-/**
- * validateOrderBy validates the 'orderBy' clause of the query.
- * Currently supports strictly one 'orderBy' condition that must match the 'where' clause key.
- *
- * @param {OrderBy<T>[] | undefined} orderBy - The 'orderBy' clause to validate.
- * @param {string} whereKey - The key used in the 'where' clause for compatibility.
- * @throws {Error} Throws an error if the 'orderBy' clause is invalid or doesn't match the 'where' clause key.
- */
 function validateOrderBy<T extends Schema>(
   orderBy: OrderBy<T>[] | undefined,
   whereKey: string,
@@ -134,13 +105,6 @@ function validateOrderBy<T extends Schema>(
   }
 }
 
-/**
- * validateSelect validates the 'select' fields of a query.
- *
- * @param {SelectField<T>[] | undefined} select - The 'select' clause to validate.
- * @param {IndexMeta[]} headers - List of headers to check for field existence.
- * @throws {Error} Throws an error if any field in the 'select' clause doesn't exist in headers.
- */
 function validateSelect<T extends Schema>(
   select: SelectField<T>[] | undefined,
   headers: IndexHeader[],
@@ -167,20 +131,10 @@ function validateSelect<T extends Schema>(
   }
 }
 
-/**
- * validateQuery checks the structure and syntax of the query against the provided headers.
- * It ensures that the fields specified in the query are valid and exist in the headers.
- * If any part of the query is invalid (e.g., a field doesn't exist), it throws an error.
- *
- * @param {Query<T>} query - The query object to validate.
- * @param {IndexMeta[]} headers - The headers against which to validate the query fields.
- * @throws {Error} Throws an error if query is invalid.
- */
 export function validateQuery<T extends Schema>(
   query: Query<T>,
   headers: IndexHeader[],
 ): void {
-  // validate the query
   validateWhere(query.where, headers);
   validateOrderBy(query.orderBy, query.where![0].key as string);
   validateSelect(query.select, headers);
