+++ src/tests/multipart.test.ts
@@ -1,4 +1,4 @@
-import parseMultipartBody from "../multipart";
+import parseMultipartBody from "../resolver/multipart";
 
 async function collect<T>(gen: AsyncGenerator<T>) {
   const result: T[] = [];
